refactor ai-billing
