+++ packages/realm-server/lib/allowed-proxy-destinations.ts
@@ -56,10 +56,6 @@ export class AllowedProxyDestinations {
     return config?.supportsStreaming ?? false;
   }
 
-  getAllowedDestinations(): string[] {
-    return Object.keys(this.destinations);
-  }
-
   static getInstance(configJson: string) {
     if (!AllowedProxyDestinations.instance) {
       AllowedProxyDestinations.instance = new AllowedProxyDestinations(
